moved api call to custom hook
